requirements.txt
python-binance>=1.0.19
websocket-client>=1.6.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0